pyproject.toml
[tool.poetry]
name = "loldrivers"
version = "0.1.0"
description = "loldrivers data generator"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
Jinja2 = "^3.1.2"
stix2 = "^3.0.1"
colorama = ">=0.3.9"
future = ">=0.16.0"
requests = ">=2.20.0"
configparser = ">=3.5.0"
pymisp = ">=2.4.123"
flask = ">=1.0"
flask-caching = "^2.0.2"
cfscrape = "^2.1.1"
pyzipper = "^0.3.6"
dnspython = "^2.3.0"
ipy = "^1.1"
valhallaapi = "^0.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
